src/requirements.txt
yt-dlp
spotdl
django>=5.2
djangorestframework>=3.14
django-cors-headers>=4.0
spotipy
ytmusicapi

